pyproject.toml
[project]
name = "googleClient"
version = "0.1.0"
description = "Google Drive domain-wide delegation admin shell (read-only)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "google-api-python-client",
  "google-auth",
  "google-auth-httplib2",
  "google-auth-oauthlib"
]

[project.scripts]
gC = "googleClient.cli:main"
